backend/requirements/development.txt
-r ./base.txt

django-extensions==3.1.5

pytest==7.2.2
pytest-cov==3.0.0
pytest-django==4.5.1
pytest-mock==3.7.0

ipdb==0.13.11
Faker==17.6.0
factory_boy==3.2.1
freezegun==1.2.2

Sphinx==5.3.0

psycopg2-binary==2.9.3
